fix: correct English language typos and improve style (#90)
It was not perfectly clear when reading the original phrasing of the following sentence. I hope that the following makes the intention more clear.

Place each page, each view model, and each service in a separate file. Apply the single responsibility principle (SRP) to all pages, all view models, all services, and other symbols. This helps make the app cleaner, easier to read and maintain, and more testable.

I converted some curly double quotation marks to normal " " to maintain the style of this particular document.