fix: correct a few typos, make minor style improvements (#91)
Most of the minor style improvements just make the text slightly easier and quicker and clearer.

On lines 52, 141, and 365 "to path" was written and I think that this is not perfectly clear. I installed NativeScript yesterday on Windows and went through this section. 

Later in the text (in the mac section, there was reference to `$PATH` which may well be the best solution in those lines as well. 

I did not many any changes to those three lines.